// tcuserapp-frontend/src/models/ScdUserModel.ts
export interface SCDUser {
  scdUserId?: number;
  country?: string;
  displayName?: string;
  fullName?: string;
  mobileNumber?: string;
  organization?: string;
  firstName?: string;
  email?: string;
  sponsor?: string;
  department?: string;
  organizationID?: string;
  lineManager?: string;
  lastName?: string;
  gid?: string;
}